Refactor training, move encoder functions into multiple files
